cleanup some unneeded code
