+++ packages/store/addon/-private/system/schema-definition-service.ts
@@ -8,7 +8,7 @@ import { HAS_MODEL_PACKAGE } from '@ember-data/private-build-infra';
 
 import type { RecordIdentifier } from '../ts-interfaces/identifier';
 import type { AttributesSchema, RelationshipsSchema } from '../ts-interfaces/record-data-schemas';
-import type Store from './ds-model-store';
+import type Store from './core-store';
 import normalizeModelName from './normalize-model-name';
 
 type ModelForMixin = (store: Store, normalizedModelName: string) => Model | null;
